refactor help bots into fewer ones
