+++ Beer Judge Trainer/DynamicLanguageBot.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class BeerRangeInfoBot {
+class DynamicLanguageBot {
 
     // BJCP SCORE RANGES
     //
@@ -58,4 +58,16 @@ class BeerRangeInfoBot {
         }
     
     }
+    
+    // function to return (language, background colors) for results page after being given a double.
+    
+    //  Point Difference Ranges
+    
+    //  0-1:  Amazing       green and sparkles
+    //  1-3:  Good          green
+    //  4-6:  Ok            yellow
+    //  6-8:  Not so close  orange
+    //  8-*:  Way off       red
+    //
+
 }
