--- Beer Judge Trainer/TrainerSuccessLevelBot.swift
@@ -1,21 +1,0 @@
-//
-//  ResultsLanguageBot.swift
-//  Beer Judge Trainer
-//
-//  Created by Nate on 3/15/17.
-//  Copyright © 2017 Nathan T. Baker. All rights reserved.
-//
-
-import Foundation
-
-// class to return (language, background colors) for results page after being given a double.
-
-//  Point Difference Ranges
-
-//  0-1:  Amazing       green and sparkles
-//  1-3:  Good          green
-//  4-6:  Ok            yellow
-//  6-8:  Not so close  orange
-//  8-*:  Way off       red
-//
-
